Move migration command to console component
